+++ backend/dashboardd/dashboardd.go
@@ -230,11 +230,13 @@ func newBackendProxy(port int, TLS *types.TLSOptions) (*httputil.ReverseProxy, e
 
 	// Configure TLS
 	if TLS != nil {
+		target.Scheme = "https"
+
 		cfg, err := TLS.ToTLSConfig()
 		if err != nil {
 			return nil, err
 		}
-		target.Scheme = "https"
+		cfg.InsecureSkipVerify = true // skip host verification on loopback interface
 		transport.TLSClientConfig = cfg
 	}