+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Added
+- Added unit test coverage for check routers.
+
 ### Changed
 - The Backend struct has been refactored to allow easier customization for the
 enterprise edition.
@@ -17,6 +20,7 @@ enterprise edition.
 - Fixed a bug in build.sh where versions for Windows and Mac OS were not
 generated correctly.
 - Display the name of extensions with table formatting in sensuctl.
+- Fixed TLS issue that occurred when dashboard communicated with API.
 - Check TTL now works with round robin checks.
 - Format string for --format flag help now shows actual arguments.